(MESS) Apollo fixes: [Hans Ostermeyer]  * fixed MLOG: machine -> machine()  * use the correct address space for dma (supplied space parameter is wrong; why?)  * minor fix for 64 Bit Ubuntu  * block_set_filemark() failed to set the filemark buffer; memcpy will now use 64 bit words (i.e. 8 byte chunks) for copying

+++ src/mess/machine/apollo_eth.c
@@ -115,7 +115,7 @@ static int apollo_eth_receive_packet(device_t *device)
 	}
 	else if (packet_len > sizeof(rx_buffer))
 	{
-		DLOG(("apollo_eth_receive_packet: data size (%d) exceeds rx buffer size (%d)!!!", packet_len, sizeof(rx_buffer)));
+		DLOG(("apollo_eth_receive_packet: data size (%d) exceeds rx buffer size (%d)!!!", packet_len, (int)sizeof(rx_buffer)));
 		return -1;
 	}
